Make httpclient dependency of H2O provided
The "httpclient" dependency is very popular and used
across many projects. In this case, H2O depends on it
and it conflicted with AWS dependencies when we use
H2O OpenML in Spark jobs on AWS EMR.

This way we make it provided so that it can be decided
by users.