PULSEDEV-26668: Support old DataRobot versions
Our DataRobot provider was assuming that Jar files contained a "classLabels" field
in the Model class, which would describe the possible target values for the target
variable being predicted.

That is not the case for old DataRobot versions. As DataRobot supports only binary
classification in exported Jars, we thus assume that the values are "0" and "1" in
that order --- i.e., the positive class is the second one, as is typical on binary
classification problems.