openml-h2o: Force schema compatibility between H2O model schema and the openml schema (#11)
* openml-h2o: Force schema compatibility between H2O model schema and the openml

Summary:
In the context of QA testing, it was detected than when attempting to import a model with a given schema that contains a subset of the total fields with which the model
was trained, the loading would succeed, while the expected behaviour would be for the model to fail.

As such, this commit adds a new validation step to make sure that the schema included in the model to be loaded is compatible with the correspondent openml schema.

* Make the validation step adaptative to a model with or without target variable

* Fix Codacy Issues