openml-datarobot: Remove javafx dependency (#12)
Summary:
Use were using the javafx Pair util class, which seems a bit overkill, since we were dependending on javafx just for this. Furthermore, javafx is not
a default java library for all environments, so environments which do not have javafx will fail when using this logic.

The javafx Pair class was replaced by the apache commons lang3 version, which does everything we want, and was already a dependency of our project.