openml-datarobot: Fix method classify() to return the index of the inferred target value.
The method should return the index of the highest value in the scores' list.
Before the change, it returns the value of the probability of the first element rounded.
In case that the first element is highest, it will return 1 which indicates that the index 1, i.e., the second element is highest.

For this reason, it is changed to the opposite, e.g., 1 - (int) Math.round(score).
Now, it returns the index 0 for the above case, and index 1 in case that the first element is lowest one.