openml-datarobot: Override UT to test Classify for H2O
The H2O uses a threshold for the classification. Therefore the method classify does not point to the index of the maximum value of the scores' list.

This was solved by overriding the method #classifyIndexOfMaxScoresValue().
The method compares the variable _defaultThreshold and the second value of the scores' list.
In this way, the classifier will point to the index of the second value in case it is bigger than _defaultThreshold, otherwise it will point to the first value.