fix PULSEDEV-31040 openml-h20: Parse Deep Learning model parameter hidden when using array format
When training a H20 Deep learning model, if the default value for the parameter
"Hidden" is used "[200,200]" or another value that follows the format of the
hint "[100,100]" the job will crash.

This happens because when this parameter is parsed the value is assumed to be
in a different format, that doesn't contain square brackets at the beginning
and end of the list of integers e.g. "200,200".

To fix this, when parsing the value, if present the square brackets are removed
from the beginning and the end of the string.