ref PULSEDEV-32373 H20 parameter validation : Added h20 parameter validation and refactored code to clean train in H2OApp
H20 Algorithm parameters were only validated in the training step (inside the trainModel method of the ModelBuilder object).
Validation occurs by creating the ModelBuilder object and then inspecting the errors to see if the errorCount is > 0.
What was done was changing the AbstractH2OParamUtils to have a validateParams method that is specific for each H2O algorithm and what it does is create the model and evaluate the errorCount. This is called in the validateForFit method of the H2OModelCreator class, that delegates to the validate method of the H2OApp.
Since both this validate method and the train method had a switch case to obtain the model depending on the algorithm, this logic was put in a factory class, H2OAlgoUtilsFactory.
Further refactoring was done to remove code duplication, and the AbstractH2OParamUtils class was changed to AbstractH2OAlgoUtils and logic was added so this class is now responsible for also providing the model construction. This makes the H2OApp way simpler since it can delegate the training or validation to the Utils class (that is now the only place in the code that is algorithm specific).
The Utils classes for each algorithm were changed to comply with the new interface and unit tests were added to test the validation of parameters.