Remove fat-jars and update dependencies
This commit removes the distribution of fat-jars with all the dependencies
because OpenML-Java is a library and thus it's better to distribute it as such,
and the project using it, can exclude some dependencies if it requires.
It also updates some dependencies, that can be safely updated without
breaking anything. It updates:
* gson to v2.9.0
* slf4j-api to v1.7.36
* commons-lang3 to v3.13.0
* log4j to v2.20.0