Refactor code and add tests
This commit changes the code structure by adding a new enum representing the libc implementation available. In order to detect the libc implementation, an env variable FDZ_OPENML_JAVA_LIBC must be set with either 'glibc' or 'musl', and using glibc if this env variable is not defined. This new enum is also used along side with the cpu architecture enum to represent the infrastructure running the code.