Add support for Alpine on LGBM (#145)
* Add support for Alpine on LGBM

This commit bumps lightgbm to v0.9.14 and makes it possible to use LGBM on nodes with AMD64 architecture using Alpine.

* Change build structure when loading dependencies

This commit changes the paths when loading the dependencies on amd64 architecture in order to reflect the changes made on the jar.

* Change build structure when loading dependencies

This commit changes the path when loading the dependencies on amd64 architecture in order to reflet the change on the jar structure from amd64/alpine to amd64/musl.

* Update make-lightgbm submodule

* Refactor code and add tests

This commit changes the code structure by adding a new enum representing the libc implementation available. In order to detect the libc implementation, an env variable FDZ_OPENML_JAVA_LIBC must be set with either 'glibc' or 'musl', and using glibc if this env variable is not defined. This new enum is also used along side with the cpu architecture enum to represent the infrastructure running the code.

* Add github action to test LightGBM on Alpine

* Try action change

* Try action change

* Try action change

* Try action change

* Try action change

* Add toString method to Infrastructure

* Add unit test for Infrastructure

* Try action change

* Try action change

* Try action change

* Try action change

* Try action change

* Try action change

* Update READEME

This commit includes information in the README about the new ENV variable used to load dependencies for musl-based OS.

* Add Override annotation to toString method on Infrastructure

Co-authored-by: Artur Pedroso <artpdr@users.noreply.github.com>

* Fix toString method on Infrastructure

* Add ENV var name as a constant

* Make FDZ_OPENML_JAVA_LIBC variable private

* Refactor Infrastructure#getLgbmNativeLibsFolder

This commits changes the if statment on the method Infrastructure#getLgbmNativeLibsFolder into a switch statement.

* Fix unknownLibcImplementationThrowsException test

* Remove unnecessary variable on getLgbmNativeLibsFolder

* Make getLibcImplementation method protected

---------

Co-authored-by: Renato Azevedo <renato.azevedo@feedzai.com>
Co-authored-by: Artur Pedroso <artur.pedroso@feedzai.com>
Co-authored-by: Artur Pedroso <artpdr@users.noreply.github.com>